==> PART 1 Introduction to Data Analysis/5_Data_Analysis_Process-Case_Study2/5_assessing.ipynb <==
# Assessing
Use the space below to explore `all_alpha_08.csv` and `all_alpha_18.csv` to answer the quiz questions below.

import pandas as pd

df1 = pd.read_csv('all_alpha_08.csv')
df1.head()

df2 = pd.read_csv('all_alpha_18.csv')
df2.head()

df1.info()

## QUESTION 1 OF 6
Find the correct count for each of the following in the 2008 dataset

df1.shape

df1.duplicated().sum()

df1.isna().any(axis=1).sum()

## QUESTION 2 OF 6
Find the correct count for each of the following in the 2018 dataset

df2.shape

df2.duplicated().sum()

df2.isna().any(axis=1).sum()

## QUESTION 3 OF 6
Match the datatype for each feature (some of these may not be ideal)

df1.info()

df2.info()

## QUESTION 4 OF 6
Match the number of non-null unique values for each of the following features

df1['SmartWay'].nunique()

df2['SmartWay'].nunique()

df2['Cert Region'].nunique()

df1['Trans'].nunique()

## QUESTION 5 OF 6
How do the Cyl columns in the 2008 and 2018 datasets differ?

df1.info()

df2.info()

## QUESTION 6 OF 6
Where are each of these fuel types present?

df1.Fuel.unique()

df2.Fuel.unique()
