==> PART 1 Introduction to Data Analysis/5_Data_Analysis_Process-Case_Study2/8_inspect_datatypes.ipynb <==
# Inspecting Data Types
Use the space below to explore `data_08_v2.csv` and `data_18_v2.csv` to answer the quiz questions below regarding datatypes. You should've created these data files in the previous section: *Filter, Drop Nulls, Dedupe*.

# load the data
import pandas as pd

df_08 =  pd.read_csv('data_08_v2.csv')
df_18 =  pd.read_csv('data_18_v2.csv')

df_08.head(1)

df_18.head(4)

df_08.dtypes

df_18.dtypes

df_08.head(20)

df_08.air_pollution_score.unique()

df_18.air_pollution_score.unique()

df_08.dtypes

df_08.head(1)

df_18.dtypes

greenhouse_gas_score
