==> PART 1 Introduction to Data Analysis/5_Data_Analysis_Process-Case_Study2/15_merging_data.ipynb <==
# Merging Datasets
Use pandas Merges to create a combined dataset from `clean_08.csv` and `clean_18.csv`. You should've created these data files in the previous section: *Fixing Data Types Pt 3*.

# load datasets
import pandas as pd

df_08 = pd.read_csv('clean_08.csv')
df_18 = pd.read_csv('clean_18.csv')

### Create combined dataset

# rename 2008 columns

df_08.rename(columns =lambda x: x[:10] + "_2008", inplace=True)

# view to check names
df_08.head()

# merge datasets
df_combined = df_08.merge(df_18, left_on='model_2008', right_on='model', how='inner')

# view to check merge
df_combined.head()

df_combined.columns

df_combined.shape

Save the combined dataset

df_combined.to_csv('combined_dataset.csv', index=False)
